api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.35
PyMySQL==1.1.1
alembic==1.13.2
marshmallow==3.21.1
gunicorn==22.0.0
flask-swagger-ui==4.11.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
python-dotenv==1.0.1
prometheus-client==0.20.0
pytest==8.3.2
pytest-flask==1.3.0
factory-boy==3.3.0
cryptography==43.0.1

